test case fixed

+++ MessengerMaven/src/main/java/messenger/Interface/GetUser.java
@@ -7,6 +7,6 @@ public interface GetUser {
 	 * @param username for which the id is to be returned
 	 * @return user-id that belongs to the username
 	 */
-	public int getUser(String username);
+	public Long getUser(String username);
 
 }
